Change quantity fields from double to int
Stock quantities and transaction quantities are inherently whole numbers. This change updates all quantity-related fields across entities, models, repositories, use cases, and presentation layer from `double` to `int`, removing unnecessary `.toDouble()` conversions.

+++ lib/core/errors/failures.dart
@@ -44,6 +44,6 @@ class InsufficientStockFailure extends AppFailure {
        );
 
   final String productId;
-  final double available;
-  final double requested;
+  final int available;
+  final int requested;
 }
